+++ pkg/statistics/handle/ddl/ddl_test.go
@@ -17,12 +17,12 @@ package ddl_test
 import (
 	"context"
 	"fmt"
+	ddlutil "github.com/pingcap/tidb/pkg/ddl/util"
 	"testing"
 
 	"github.com/pingcap/tidb/pkg/meta/model"
 	pmodel "github.com/pingcap/tidb/pkg/parser/model"
 	"github.com/pingcap/tidb/pkg/planner/cardinality"
-	"github.com/pingcap/tidb/pkg/statistics/handle/util"
 	"github.com/pingcap/tidb/pkg/testkit"
 	"github.com/pingcap/tidb/pkg/types"
 	"github.com/pingcap/tidb/pkg/util/mock"
@@ -1302,13 +1302,10 @@ func TestAddPartitioning(t *testing.T) {
 	)
 }
 
-func findEvent(eventCh <-chan *util.DDLEvent, eventType model.ActionType) *util.DDLEvent {
+func findEvent(eventCh <-chan *ddlutil.SchemaChangeEvent, eventType model.ActionType) *ddlutil.SchemaChangeEvent {
 	// Find the target event.
 	for {
 		event := <-eventCh
-		if event.SchemaChangeEvent.GetType() == eventType {
-			return event
-		}
 		if event.GetType() == eventType {
 			return event
 		}
